+++ src/components/duelling-picklist/picklist-item/picklist-item.spec.tsx
@@ -4,7 +4,7 @@ import { mount } from "enzyme";
 import PicklistItem, { PicklistItemProps } from "./picklist-item.component";
 import { StyledButton, StyledPicklistItem } from "./picklist-item.style";
 import StyledIcon from "../../icon/icon.style";
-import FocusContext from "../duelling-picklist.context";
+import FocusContext from "../__internal__/duelling-picklist.context";
 import { assertStyleMatch } from "../../../__spec_helper__/__internal__/test-utils";
 import CarbonProvider from "../../carbon-provider/carbon-provider.component";
