+++ src/components/duelling-picklist/picklist/picklist.component.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useContext, useEffect, useMemo } from "react";
 import { TransitionGroup } from "react-transition-group";
 
 import { StyledPicklist, StyledEmptyContainer } from "./picklist.style";
-import FocusContext from "../duelling-picklist.context";
+import FocusContext from "../__internal__/duelling-picklist.context";
 import Events from "../../../__internal__/utils/helpers/events";
 import PicklistGroup from "../picklist-group/picklist-group.component";
